requirements.txt
absl-py
urlfile
git+https://github.com/fusepy/fusepy